Make Notifications::Fanout#listeners_for faster
Previously we stored all subscribers in an array, and every time a new
message name came in asked each subscriber if they responded to the
message.

This commit changes Fanout to hash subscribers with a String pattern by
their pattern. This way we can look them up directly and only do the
O(N) scan over the non-string (Regex or any) patterns.